pyproject.toml
[tool.poetry]
name = "python-advanced-diploma"
version = "0.1.0"
description = "clone twitter"
authors = ["evgenator <zenyavolobuev@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.114.0"
python-dotenv = "^1.0.1"
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.34"}
alembic = "^1.13.2"
asyncpg = "^0.29.0"
uvicorn = "^0.30.6"
jinja2 = "^3.1.4"
python-multipart = "^0.0.9"
aiofiles = "^24.1.0"
types-aiofiles = "^24.1.0.20240626"


[tool.poetry.group.dev.dependencies]
httpx = "^0.27.2"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.3.2"
mypy = "^1.11.2"
black = "^24.8.0"
isort = "^5.13.2"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
